fix_comments (#3)
Generally, there is a lot of newly added documentation for the scripts. This way, in the future, the code will be able to be understood and maintained better.

The README.md was also updated with new, relevant information that hopefully will make using the code easier. Further, there was an additional potential fix that was created to potentially help users solve the problems on their own.

There was also a minor change, albeit in the wrong branch, that removed a variable that did not need to exist; the variable was a leftover from an older version.

Co-authored-by: sw-arkerley <sw-arkerley@login.hpc.tntech.edu>
Reviewed-on: https://git.tntech.edu/ITS-RCD/md-scaling-study/pulls/3