feature_setupexits (#4)
Adds exits to the setup script to avoid errors from being created by the script trying to perform other scripts it cannot do. There are error messages added that attempt to predict what could be causing the errors in the future.

In addition, this adds an additional message that explains to the user that each analysis tool folder is being successfully created.

Co-authored-by: sw-arkerley <sw-arkerley@login.hpc.tntech.edu>
Reviewed-on: https://git.tntech.edu/ITS-RCD/md-scaling-study/pulls/4