feature_scalabilityimprovements (#5)
The most prominent change was the addition of two new functions that split up figureCreator more into a more readable form. The code splits the if statement into functions to cut down on the previously large function.
The code refactoring moved variables and functions around to get the least amount of repeated code within the two new functions.

The other change was a check of the comments that had TODOs, revealing that the sorts were relevant and one of the formats was unneeded.

Co-authored-by: sw-arkerley <sw-arkerley@login.hpc.tntech.edu>
Reviewed-on: https://git.tntech.edu/ITS-RCD/md-scaling-study/pulls/5