refactor_pycodestyle (#6)
Refactors a fair amount of the code, predominantly variable names, to have PEP 8 compliant names; file names were also modified to be of proper PEP 8 compliance. By extension, the comments were made to be compliant and modified to reflect new changes to the variable names.

The README.md was also modified to represent the change of variable and file names.

Fixes the Extractors to also be able to use the new variable names.

Co-authored-by: sw-arkerley <sw-arkerley@login.hpc.tntech.edu>
Reviewed-on: https://git.tntech.edu/ITS-RCD/md-scaling-study/pulls/6