update_SaferCode (#8)
This update focused on making the code able to handle errors that are known to cause issues. Rather than simply crashing or ignoring problems without informing the user as to why, there is now code that will elaborate and explain errors to try to make the user aware of the issue.

The main additions are:
 - the extractors no longer will attempt to load and run on empty files, creating empty csvs for that reason
 - the scalabilityplotter now has several potential errors it can catch
 - explains the errors in a short message that should convey enough to the user to know there is a problem
 - the scalabilityplotter should no longer generate empty graphs with no data
 - makes the MASTERs check if the setup script has been run before
 - changes the MASTERs to have lower times on the biggest jobs

Co-authored-by: sw-arkerley <sw-arkerley@login.hpc.tntech.edu>
Reviewed-on: https://git.tntech.edu/ITS-RCD/md-scaling-study/pulls/8