Adds a better, more dynamic method of style usage (#11)
This was committed and pushed to one branch, but it appears that branch either was never on the repository or some other issue. Regardless, this is the change it meant to implement.

All it should do is add the enough unique styles for each potential graph of various atom amounts or type of solver.

I thought this was already merged, but I am clearly mistaken.
Fixes issue #10

Reviewed-on: https://git.tntech.edu/ITS-RCD/md-scaling-study/pulls/11